+++ cli/src/config.rs
@@ -24,6 +24,8 @@ use jj_lib::settings::ConfigResultExt as _;
 use thiserror::Error;
 use tracing::instrument;
 
+use crate::command_error::{user_error, user_error_with_message, CommandError};
+
 #[derive(Error, Debug)]
 pub enum ConfigError {
     #[error(transparent)]
@@ -432,6 +434,72 @@ fn read_config_path(config_path: &Path) -> Result<config::Config, config::Config
             )
         })
         .build()
+}
+
+pub fn write_config_value_to_file(
+    key: &str,
+    value_str: &str,
+    path: &Path,
+) -> Result<(), CommandError> {
+    // Read config
+    let config_toml = std::fs::read_to_string(path).or_else(|err| {
+        match err.kind() {
+            // If config doesn't exist yet, read as empty and we'll write one.
+            std::io::ErrorKind::NotFound => Ok("".to_string()),
+            _ => Err(user_error_with_message(
+                format!("Failed to read file {path}", path = path.display()),
+                err,
+            )),
+        }
+    })?;
+    let mut doc: toml_edit::Document = config_toml.parse().map_err(|err| {
+        user_error_with_message(
+            format!("Failed to parse file {path}", path = path.display()),
+            err,
+        )
+    })?;
+
+    // Apply config value
+    // Interpret value as string if it can't be parsed as a TOML value.
+    // TODO(#531): Infer types based on schema (w/ --type arg to override).
+    let item = match value_str.parse() {
+        Ok(value) => toml_edit::Item::Value(value),
+        _ => toml_edit::value(value_str),
+    };
+    let mut target_table = doc.as_table_mut();
+    let mut key_parts_iter = key.split('.');
+    // Note: split guarantees at least one item.
+    let last_key_part = key_parts_iter.next_back().unwrap();
+    for key_part in key_parts_iter {
+        target_table = target_table
+            .entry(key_part)
+            .or_insert_with(|| toml_edit::Item::Table(toml_edit::Table::new()))
+            .as_table_mut()
+            .ok_or_else(|| {
+                user_error(format!(
+                    "Failed to set {key}: would overwrite non-table value with parent table"
+                ))
+            })?;
+    }
+    // Error out if overwriting non-scalar value for key (table or array) with
+    // scalar.
+    match target_table.get(last_key_part) {
+        None | Some(toml_edit::Item::None | toml_edit::Item::Value(_)) => {}
+        Some(toml_edit::Item::Table(_) | toml_edit::Item::ArrayOfTables(_)) => {
+            return Err(user_error(format!(
+                "Failed to set {key}: would overwrite entire table"
+            )));
+        }
+    }
+    target_table[last_key_part] = item;
+
+    // Write config back
+    std::fs::write(path, doc.to_string()).map_err(|err| {
+        user_error_with_message(
+            format!("Failed to write file {path}", path = path.display()),
+            err,
+        )
+    })
 }
 
 /// Command name and arguments specified by config.
